requirements.txt
requests==2.32.5
python-telegram-bot==13.0
python-dotenv==1.2.2